requirements.txt
# Cross-Channel AI Agents Requirements

# Core Dependencies
asyncio
aiohttp==3.9.1
aiohttp-cors==0.7.0
aiohttp-jinja2==1.5.1
python-dotenv==1.0.0

# AI/ML Dependencies
openai==1.3.0
pandas==2.1.4

# Twilio SDK
twilio==8.10.0

# Web Framework (for webhook servers)
flask==3.0.0

# Templating and UI
jinja2==3.1.2
colorama==0.4.6

# Environment Configuration
pydantic==2.5.0

# Development Dependencies (optional)
pytest==7.4.3
black==23.11.0
flake8==6.1.0

# Data Processing
jsonpath-ng==1.6.1